REDCap.jl

REDCap.jl

Overview

Index

Examples

Overview

A Julia frontend for REDCap's API.

REDCap.REDCap — Module

REDCap API methods are defined as named Julia functions. The named arguments of the Julia functions each correspond to the API parameters. REDCap API parameters seem to always have camelCase names, and parameters that share a name generally have the same behavior and accepted values.

REDCap API calls generally include a content and action parameter.

In REDCap.jl, these are managed internally and don't need to be passed as an argument, since the function names generally correspond to a combination of these (e.g., delete_records() passes content=record and action=delete).

There's generally no REDCap action parameter for import or export. Instead, API calls without an action parameter are interpreted as import requests if there is a data parameter, and an export request if there is none. To preserve the expected behaviour of named functions, REDCap.jl import functions require a data argument, and export functions do not accept one.

Function docstrings indicate how to use the Julia functions. Refer to the official documentation for detailed specifications on the corresponding REDCap methods.

Syntax

Each REDCap method accepts a number of parameters that follow a shared naming convention. REDCap.jl is designed to closely follow the design and syntax patterns of REDCap. Every REDCap API method is available as a function that supplies certain required parameters and checks user inputs for validity.

Return values and REDCap messages are returned as Strings directly, but the documentation shows how these can be parsed in useful ways.

Function arguments are named after REDCap method parameters. These are passed as named arguments and take values with intuitive types.

token and url

Nearly all REDCap methods accept a token that is unique to the project and user.

The URL must exactly match this example: https://example.example/redcap/api/

Your REDCap token and your institution's REDCap API URL can be read by default from Julia's environment variables. You can make them avaiable to REDCap.jl by putting the following lines in your local Julia startup file (probably ~/.julia/config/startup.jl):

ENV["REDCAP_API_TOKEN"] = "C0FFEEC0AC0AC0DEC0FFEEC0AC0AC0DE"
ENV["REDCAP_API_URL"] = "http://example.com/redcap/api/"

They can also be passed as ordinary named arguments.

A few methods accept a super token, including create_project, which can be used to generate a project and project-level token. If you have a super token, you might wish to keep that in your startup file, generating and saving project-level tokens as needed.

data

The data parameter contains a list of attributes, which varies between REDCap methods. For definitive attribute lists, see the official REDCap documentation. These attributes generally have camel_case names.

In REDCap.jl, the corresponding data argument is implemented as a Collection. This can be a NamedTuple or a Dict, as well as a file handle or String. If you use a NamedTuple, it will be translated internally into whatever format you use (xml by default).

import_project_info(
    data=(
        project_title="New name",
        project_notes="New notes"
    ),
    returnFormat=:csv,
)

But please keep in mind that a NamedTuple with one value must contain a comma:

import_project_info(
    data=(
        project_title="New name", # this comma is required
    ),
    returnFormat=:csv,
)

A Dict value is fine as well.

import_project_info(data=Dict(:project_title=>"New name"), returnFormat=:csv)

String values are accepted. If the string is a file name, the contents of the file are sent; otherwise, the value is sent directly as part of the API request.

data_string = """
    [{"data_access_group_name":"CA Site","unique_group_name":"ca_site"},
    {"data_access_group_name":"FL Site","unique_group_name":"fl_site"},
    {"data_access_group_name":"New Site","unique_group_name":""}]
"""
out = open("data_file.json","w"); write(out, data_string); close(out)

import_DAGs(token=t,data=data_string, format=:json) # string is passed to the API

import_DAGs(token=t, data="data_file.json", format=:json) # string is pattern-matched as a filename 

As for collections, only collections of scalar entries are currently supported. So, a list of attributes and values is accepted, but a Dict containing multiple rows per column can only be read in from a file.

In the REDCap API, the presence of a data parameter often changes the behavior of a method. For instance, most import methods are implemented as an export method with an added data parameter. In REDCap.jl, it would be considered a bug for import_project_data to ever act as export_project_data, so the data paramater is almost always required where it is present.

format and returnFormat

Supported options are :csv, :json, :xml (the default value), and sometimes :odm. These values can be passed as Strings or Symbols.

For import methods, the format parameter designates user input and the returnFormat parameter applies to REDCap messages and return values. Otherwise, there is generally only a single format parameter that applies to REDCap messages and return values.

content and action

The content and action parameters are what define each REDCap method, for the most part. In REDCap.jl, these are passed internally and don't need to be supplied by the user. Instead, they're fixed for each function.

Troubleshooting

If a function call doesn't produce the expected results, try making debug messages visible for this package by running ENV["JULIA_DEBUG"] = REDCap.

The data parameter is converted to a formatted string, so you might try different format parameters (:csv, :json, or :xml).

If you're uncertain about the format for an input data parameter, try setting up an example in the browser and exporting the data. You may have to remove the values for certain generated columns before importing.

Feel free to create an issue for any unexpected errors, or for feature requests.

Index

REDCap.REDCap

REDCap.create_folder

REDCap.create_project

REDCap.delete_DAGs

REDCap.delete_arms

REDCap.delete_events

REDCap.delete_file

REDCap.delete_file_from_file_repository

REDCap.delete_records

REDCap.delete_user_roles

REDCap.delete_users

REDCap.export_DAGs

REDCap.export_PDF

REDCap.export_arms

REDCap.export_events

REDCap.export_file

REDCap.export_file_from_file_repository

REDCap.export_instrument_event_mappings

REDCap.export_instruments

REDCap.export_list_of_export_field_names

REDCap.export_list_of_folders

REDCap.export_logging

REDCap.export_metadata

REDCap.export_project_XML

REDCap.export_project_info

REDCap.export_records

REDCap.export_repeating_instruments_and_events

REDCap.export_reports

REDCap.export_survey_link

REDCap.export_survey_participants

REDCap.export_survey_queue_link

REDCap.export_survey_return_code

REDCap.export_user_DAG_assignment

REDCap.export_user_role_assignments

REDCap.export_user_roles

REDCap.export_users

REDCap.export_version

REDCap.generate_next_record_name

REDCap.import_DAGs

REDCap.import_arms

REDCap.import_events

REDCap.import_file

REDCap.import_file_from_file_repository

REDCap.import_instrument_event_mappings

REDCap.import_metadata

REDCap.import_project_info

REDCap.import_records

REDCap.import_repeating_instruments_and_events

REDCap.import_user_DAG_assignment

REDCap.import_user_role_assignments

REDCap.import_user_roles

REDCap.import_users

REDCap.rename_record

REDCap.switch_DAG