REDCap.create_folder — Method

function delete_DAGs(; url=get_url(), token=get_token(), dags)

Create a New Folder in the File Repository

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

name: name of the new folder (150 characters maximum)

folder_id: folder_id of the intended parent directory (default is top-level directory in File Repository)

dag_id: optionally restrict folder access to a Data Access Group

role_id: optionally restrict folder access to a User Role

REDCap.create_project — Method

function create_project(; data, url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, odm=nothing,)

Create a new REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict. Availabel attributes are project_title, purpose, purpose_other, project_notes, is_longitudinal, surveys_enabled, and record_autonumbering_enabled.

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.delete_DAGs — Method

function delete_DAGs(; url=get_url(), token=get_token(), dags)

Delete Data Access Groups (DAGs) from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

dags: group names (can be scalar or vector)

REDCap.delete_arms — Method

function delete_arms(; url=get_url(), token=get_token(), arms,)

Delete study arms from a REDCap project

Notes

If no value is provided for arms, all study arms are returned. Deleting a study arm deletes any included records and data.

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

arms: names of study arms (can be scalar or vector)

REDCap.delete_events — Method

function delete_events(; url=get_url(), token=get_token(), events=nothing,)

Delete Events from a longitudinal REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

events: unique event names (can be scalar or vector)

REDCap.delete_file — Method

function delete_file(; url=get_url(), token=get_token(), returnFormat=nothing, record=nothing, field=nothing, event=nothing, repeat_instance=nothing,)

Delete a file from a record in a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

record: record ID

field: name of field that contains file

event: unique event name (longitudinal projects)

repeat_instance: instance number of repeating event or instrument (where applicable, default is 1)

REDCap.delete_file_from_file_repository — Method

function delete_file_from_file_repository(; url=get_url(), token=get_token(), doc_id,)

Delete a File from the File Repository

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

doc_id: the file's doc_id

REDCap.delete_records — Method

function delete_records(; url=get_url(), token=get_token(), records, arm=nothing, instrument=nothing, event=nothing, repeat_instance=nothing,)

Delete records from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

records: record names (can be scalar or vector)

arm: optionlly limit deletions to this arm number

instrument: optionlly limit deletions to this unique instrument name

event: optionlly limit deletions to this unique event name (required for longitudinal projects, otherwise unavailable)

repeat_instance: optionlly limit deletions to this repeating instrument or event

delete_logging: if true, delete logging activity for deleted record (default is false)

REDCap.delete_user_roles — Method

function delete_user_roles(; url=get_url(), token=get_token(), roles,)

Delete roles from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

roles: unique role names (can be scalar or vector)

REDCap.delete_users — Method

function delete_users(; url=get_url(), token=get_token(), users=users)

Delete users from a REDCap project (must have delete records privilege)

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

users: user names (can be scalar or vector)

Example

julia > delete_users(users = [:user1,  :user2])

REDCap.export_DAGs — Method

function export_DAGs(; url=get_url(), token=get_token(), format=nothing,)

Export Data Access Groups (DAGs) from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_PDF — Method

function export_PDF(; url=get_url(), token=get_token(), format=nothing, record=nothing, event=nothing, instrument=nothing, repeat_instance=nothing, allRecords=nothing, compactDisplay=nothing,)

Export PDF of Data Collection Instruments

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

`record: record ID (default is a PDF with no data)

event: unique event name for a longitudinal project

instrument: unique instrument name

repeat_instance: repeat instance number for projects with repeating instruments/events

allRecords: if passed with any value, all records are exported

compactDisplay: if true, excludes empty fields and unselected options (default is false)

REDCap.export_arms — Method

function export_arms(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, arms=nothing,)

Export a REDCap project's study arms

Notes

If no value is provided for arms, all study arms are returned.

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

arms: names of study arms (can be scalar or vector)

REDCap.export_events — Method

function export_events(; url=get_url(), token=get_token(), format=nothing, arms=nothing,)

Export Events from a longitudinal REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

arms: study arms from which to pull events (all are inclued by default)

REDCap.export_file — Method

function export_file(; url=get_url(), token=get_token(), returnFormat=nothing, record=nothing, field=nothing, event=nothing, repeat_instance=nothing,)

Export a file from a record in a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

record: record ID

field: name of field that contains file

event: unique event name (longitudinal projects)

repeat_instance: instance number of repeating event or instrument (where applicable, default is 1)

REDCap.export_file_from_file_repository — Method

function export_file_from_file_repository(; url=get_url(), token=get_token(), doc_id=nothing,

Export a File from the File Repository

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

doc_id: the file's doc_id

REDCap.export_instrument_event_mappings — Method

function export_instrument_event_mappings(; url=get_url(), token=get_token(), format=nothing, arms=nothing,)

Export mappings of data collection instruments onto designated Events for a longitudinal REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

arms: study arms from which to pull events (all are inclued by default)

REDCap.export_instruments — Method

function export_instruments(; url=get_url(), token=get_token(), format=nothing,)

Export Instruments (Data Entry Forms)

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_list_of_export_field_names — Method

function export_list_of_export_field_names(; url=get_url(), token=get_token(), format=nothing, field=nothing)

Export Events from a longitudinal REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

field: field variable names (can be scalar or vector). By default, all fields are returned.

REDCap.export_list_of_folders — Method

function export_list_of_folders(; url=get_url(), token=get_token(), folder_id=nothing,)

Export a list of files for all subfolders

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

folder_id: folder_id of intended folder (default is the top-level directory of the File Repository)

REDCap.export_logging — Method

function export_logging(; url=get_url(), token=get_token(), format=nothing, logtype=nothing, user=nothing, record=nothing, dag=nothing, beginTime=nothing, endTime=nothing,)

Export mappings of data collection instruments onto designated Events for a longitudinal REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

logtype: optionally limit output to an event type (export, manage, user, record, record_add, record_edit, record_delete, lock_record, page_view)

user: optionally limit output to a specific user

record: optionally limit output to a specific record

dag: optionally limit output to a specific Data Access Group (DAG)

beginTime: optionally limit output to events logged after a given datetime

endTime: optionally limit output to events logged before a given datetime

REDCap.export_metadata — Method

function export_metadata(; url=get_url(), token=get_token(), format=nothing, fields=nothing, forms=nothing,)

Export metadata (Data Dictionary) from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

fields: field names (can be scalar or vector). By default, all fields are pulled.

fields: unique form names for data collection instruments (can be scalar or vector). By default, all fields are pulled.

REDCap.export_project_XML — Method

function export_project_XML(; url=get_url(), token=get_token(), returnFormat=nothing, returnMetadataOnly=nothing, records=nothing, fields=nothing, events=nothing, exportSurveyFields=nothing, exportDataAccessGroups=nothing, filterLogic=nothing, exportFiles=nothing,)

Export an entire REDCap project as an XML project file, which can be used to create a new project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

records: optionally limit output to these record names (can be scalar or vector)

fields: optionally limit output to these field names (can be scalar or vector)

events: optionally limit output to these unique event names (can be scalar or vector)

returnMetadataOnly: optionally return metadata with no data (default is false)

exportSurveyFields: optionally export survey identifier and timestamp fields (default is false)

exportDataAccessGroups: optionally export redcap_data_access_group field (default is false)

exportFiles: optionally include contents of File Upload and Signature fields (default is false)

filterLogic: logic string for filtering records

REDCap.export_project_info — Method

function export_project_info(; url=get_url(), token=get_token(), format=nothing,)

Export basic attributes of a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_records — Method

function export_records(; url=get_url(), token=get_token(), format=nothing, type=nothing, records=nothing, fields=nothing, forms=nothing, events=nothing, rawOrLabel=nothing, rawOrLabelHeaders=nothing, exportCheckboxLabel=nothing, exportSurveyFields=nothing, exportDataAccessGroups=nothing, filterLogic=nothing, dateRangeBegin=nothing, dateRangeEnd=nothing, csvDelimiter=nothing, decimalCharacter=nothing, exportBlankForGrayFormStatus=nothing)

Export records from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, :xml (default), or :odm

type: :flat (default) or :eav

records: optionally limit exports to specific records (can be scalar or vector)

fields: optionally limit exports to specific field names (can be scalar or vector)

forms: optionally limit exports to specific form names (can be scalar or vector)

events: (longitudinal projects only) optionally limit exports to specific unique event names (can be scalar or vector)

rawOrLabel: :raw (default, export raw coded values) or :label (export labels)

rawOrLabelHeaders: (for :flat type, :csv format only) raw or label CSV headers

exportCheckboxLabel: toggle format of checkbox field values (false by default)

exportSurveyFields: optionally export survey identifier and timestamp fields (false by default)

exportDataAccessGroups: optionally export the redcap_data_access_group field (false by default)

filterLogic : optionally filter records based on a logic string

dateRangeBegin: optionally limit output to records created or modified after a given datetime

dateRangeEnd: optionally limit output to records created or modified before a given datetime

csvDelimiter: ',' (default), 'tab', ';', '|', or '^'

decimalCharacter: enforce ',' or '.' format (empty by default)

exportBlankForGrayFormStatus: optionally export blank values for instrument complete status fields with a gray status icon

REDCap.export_repeating_instruments_and_events — Method

function export_repeating_instruments_and_events(; url=get_url(), token=get_token(), format=nothing,)

Export a list of repeating instruments and events for a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_reports — Method

function export_reports(; url=get_url(), token=get_token(), report_id, format=nothing, returnFormat=nothing, fields=nothing, rawOrLabel=nothing, rawOrLabelHeaders=nothing, exportCheckboxLabel=nothing, csvDelimiter=nothing, decimalCharacter=nothing,)

Export a report's data set for a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

report_id: the report ID number

format: the desired output format: :csv, :json, :xml (default), or :odm

rawOrLabel: :raw (default, export raw coded values) or :label (export labels)

rawOrLabelHeaders: (for :flat type, :csv format only) raw or label CSV headers

exportCheckboxLabel: toggle format of checkbox field values (false by default)

csvDelimiter: ',' (default), 'tab', ';', '|', or '^'

decimalCharacter: enforce ',' or '.' format (empty by default)

REDCap.export_survey_link — Method

function export_survey_link(; url=get_url(), token=get_token(), returnFormat=nothing, record=nothing, instrument=nothing, event=nothing, repeat_instance=nothing,)

Export a participant's survey link for a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

record: the record name

instrument: the instrument's name (must be enabled as a survey)

event: the unique event name (longitudinal projects only)

repeat_instance: the repeat instance of the event or instrument (default is 1)

REDCap.export_survey_participants — Method

function export_survey_participants(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, instrument=nothing, event=nothing, repeat_instance=nothing,)

Export a list of survey participants

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

instrument: the instrument's name (must be enabled as a survey)

event: the unique event name (longitudinal projects only)

REDCap.export_survey_queue_link — Method

function export_survey_queue_link(;record=nothing,returnFormat=nothing)

Export a participant's survey queue link

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

record: the record name

REDCap.export_survey_return_code — Method

function export_survey_link(; url=get_url(), token=get_token(), returnFormat=nothing, record=nothing, instrument=nothing, event=nothing, repeat_instance=nothing,)

Export a participant's survey return code

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

record: the record name

instrument: the instrument's name (must be enabled as a survey)

event: the unique event name (longitudinal projects only)

repeat_instance: the repeat instance of the event or instrument (default is 1)

REDCap.export_user_DAG_assignment — Method

function export_user_DAG_assignment(; url=get_url(), token=get_token(), format=nothing,)

Export users' assignments to Data Access Groups (DAGs) for a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_user_role_assignments — Method

function export_user_role_assignments(; url=get_url(), token=get_token(), format=nothing, )

Export role assignments from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_user_roles — Method

function export_user_roles(; url=get_url(), token=get_token(), format=nothing,)

Export roles from a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_users — Method

function export_users(; url=get_url(), token=get_token(), format=nothing,)

Export users and user privileges for a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.export_version — Method

function export_version(; url=get_url(), token=get_token(),)

Export the REDCap version

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

REDCap.generate_next_record_name — Method

function generate_next_record_name(; url=get_url(), token=get_token(),)

Generate the next record name

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

REDCap.import_DAGs — Method

function import_DAGs(; url=get_url(), token=get_token(), data, format=nothing, returnFormat=nothing,)

Export users' assignments to Data Access Groups (DAGs) for a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_arms — Method

function import_arms(; url=get_url(), token=get_token(), format=nothing, data, returnFormat=nothing, override=nothing,)

Import new study arms to a REDCap project, or rename existing arms

Notes

Deleting a study arm deletes any included records and data.

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

arms: names of study arms (can be scalar or vector)

override: if true, all existing arms are erased; if false (default), existing arms can only be renamed

data: May be a String, a file name, or a data type such as NamedTuple or Dict

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_events — Method

function import_events(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, data=nothing, override=nothing,)

Import new Events to a longitudinal REDCap project, or update the attributes of existing Events

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict

override: if true, all existing Events are erased; if false (default), existing Events can only be updated

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_file — Method

function import_file(; url=get_url(), token=get_token(), returnFormat=nothing, record=nothing, field=nothing, event=nothing, repeat_instance=nothing, file=nothing,)

Export a file from a record in a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

returnFormat: the desired output format: :csv, :json, or :xml (default)

record: record ID

field: name of field that contains file

event: unique event name (longitudinal projects)

repeat_instance: instance number of repeating event or instrument (where applicable, default is 1)

file: file contents to be uploaded

REDCap.import_file_from_file_repository — Method

function importfilefromfilerepository(; url=geturl(), token=gettoken(), format=nothing, file, folder_id=nothing,)

Export a list of files for all subfolders function importfilefromfilerepository(; url=geturl(), token=gettoken(), format=nothing, returnFormat=nothing, name=nothing, file=nothing, folder_id=nothing,)

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

file: contents of the new file

folder_id: folder_id of intended folder (default is the top-level directory of the File Repository)

format: the desired output format: :csv, :json, or :xml (default)

REDCap.import_instrument_event_mappings — Method

function import_instrument_event_mappings(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, data,)

Import mappings of data collection instruments onto designated Events for a longitudinal REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_metadata — Method

function importmetadata(; data, url=geturl(), token=get_token(), format=nothing, returnFormat=nothing,)

Import metadata (Data Dictionary) to a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_project_info — Method

function import_project_info(; format=nothing, data, url=get_url(), token=get_token(),)

Import basic attributes of a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String,a file name,or a data type such as NamedTuple or Dict. Available attributes are project_title, project_language, purpose, purpose_other, project_notes, custom_record_label, secondary_unique_field, is_longitudinal, surveys_enabled, scheduling_enabled, record_autonumbering_enabled, randomization_enabled, project_irb_number, project_grant_number, project_pi_firstname, project_pi_lastname, display_today_now_button, and bypassbranchingerasefieldprompt`.

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

REDCap.import_records — Method

function import_records(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, type=nothing, overwriteBehavior=nothing, forceAutoNumber=nothing, backgroundProcess=nothing, data, dateFormat=nothing, csvDelimiter=nothing, returnContent=nothing,)

Import records to a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

format: the desired output format: :csv, :json, :xml (default), or :odm

type: :flat (default) or :eav

overwriteBehavior: overwrite data with blank values (by default, these are ignored)

forceAutoNumber: optionally determine record numbers automatically (these must be provided by default)

backgroundProcess: set to true for large uploads

data: May be a String, a file name, or a data type such as NamedTuple or Dict

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

dateFormat: :MDY, :DMY, or :YMD (default)

csvDelimiter: ',' (default), 'tab', ';', '|', or '^'

returnContent: sets the return value (:count (default, :lds, or :auto_lds)

REDCap.import_repeating_instruments_and_events — Method

function import_repeating_instruments_and_events(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, data,)

Import a list of repeating instruments and events to a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_user_DAG_assignment — Method

function import_user_DAG_assignment(; url=get_url(), token=get_token(), data, format=nothing, returnFormat=nothing,)

Export users' assignments to Data Access Groups (DAGs) for a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_user_role_assignments — Method

function import_user_role_assignments(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, data,)

Import role assignments to a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict. Available attributes are user_name, unique_role_name, and data_access_group.

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_user_roles — Method

function import_user_roles(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, data,)

Import roles to a REDCap project

An important REDCap API note: unique_role_name is generated by REDCap and must be left blank. You have to clear the uniquerolename column to import the roles you've exported from a project.

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict. Available attributes are unique_role_name, role_label, design, alerts, user_rights, data_access_groups, reports, stats_and_charts, manage_survey_participants, calendar, data_import_tool, data_comparison_tool, logging, email_logging, file_repository, data_quality_create, data_quality_execute, api_export, api_import, api_modules, mobile_app, mobile_app_download_data, record_create, record_rename, record_delete, lock_records_customization, lock_records, lock_records_all_forms, forms, and forms_export.

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.import_users — Method

function import_users(; url=get_url(), token=get_token(), format=nothing, returnFormat=nothing, data,)

Import users and user privileges to a REDCap project

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

data: May be a String, a file name, or a data type such as NamedTuple or Dict. Available attributes are username, expiration, dataaccessgroup, design, alerts, userrights, dataaccessgroups, dataexport, reports, statsandcharts, managesurveyparticipants, calendar, dataimporttool, datacomparisontool, logging, emaillogging, filerepository, dataqualitycreate, dataqualityexecute, apiexport, apiimport, apimodules, mobileapp, mobileappdownloaddata, recordcreate, recordrename, recorddelete, lockrecordscustomization, lockrecords, lockrecordsallforms, forms, and forms_export.

format: the format of the data input parameter: :csv, :json, or :xml (default). If data is a String or a file name, this value must indicate the correct format. If data is a NamedTuple, Dict, or similar type, this value will determine what format will be used internally to pass on the data.

returnFormat: the desired output format: :csv, :json, or :xml (default)

REDCap.rename_record — Method

function rename_record(; url=get_url(), token=get_token(), record, new_record_name, arm=nothing,)

Rename a record

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

record: current record name

new_record_name: new record name

arm: optionally limit renaming to this arm number

REDCap.switch_DAG — Method

function import_user_DAG_assignment(; url=get_url(), token=get_token(), data, format=nothing, returnFormat=nothing,)

Switch the REDCap API user's current Data Access Group (DAG)

Named arguments

url: (read from ENV["REDCAP_API_URL"] by default)

token: an API token specific to the REDCap project and username (read from ENV["REDCAP_API_TOKEN"] by default)

dag: unique group name for the intended DAG